media/other/module8/theme2/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.6
Django==3.2
djangorestframework==3.12.4
flake8==7.1.1
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.4.3
toml==0.10.2
